Move all Repairable logic except RepairActors to new ResupplyManager trait
