+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -471,6 +471,7 @@
     <Compile Include="Traits\Render\DrawLineToTarget.cs" />
     <Compile Include="Traits\RejectsOrders.cs" />
     <Compile Include="Traits\Repairable.cs" />
+    <Compile Include="Traits\ResupplyManager.cs" />
     <Compile Include="Traits\RepairsBridges.cs" />
     <Compile Include="Traits\SeedsResource.cs" />
     <Compile Include="Traits\StoresResources.cs" />
