+++ mods/d2k/rules/defaults.yaml
@@ -191,7 +191,8 @@
 	HiddenUnderFog:
 	ActorLostNotification:
 	Repairable:
-		RepairBuildings: repair_pad
+		RepairActors: repair_pad
+	ResupplyManager:
 	Guard:
 		Voice: Guard
 	Guardable:
